requirements.txt
anyio==4.10.0
azure-ai-inference==1.0.0b9
azure-core==1.35.0
certifi==2025.8.3
charset-normalizer==3.4.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
python-dotenv==1.1.1
python-telegram-bot==22.3
requests==2.32.5
six==1.17.0
sniffio==1.3.1
typing_extensions==4.15.0
urllib3==2.5.0
